requirements.txt
anyascii==0.3.3
asgiref==3.11.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-modelcluster==6.4.1
django-permissionedforms==0.1
django-stubs-ext==5.2.9
django-taggit==6.1.0
django-tasks==0.11.0
django-treebeard==4.8.0
djangorestframework==3.16.1
draftjs_exporter==5.2.0
drf-spectacular==0.29.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
laces==0.1.2
modelsearch==1.2
pillow==12.1.1
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
soupsieve==2.8.3
sqlparse==0.5.5
telepath==0.3.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wagtail==7.3.1
Willow==1.12.0
